requirements.txt
psycopg2==2.9.1
django==3.2.6
django-crispy-forms==1.12.0
easy-thumbnails==2.7.1
django-cleanup==5.2.0
django-simple-captcha==0.5.14
djangorestframework==3.12.4
